+++ light-curve-python/src/features.rs
@@ -4,15 +4,13 @@ use crate::np_array::{Arr, GenericFloatArray1};
 use crate::sorted::is_sorted;
 
 use const_format::formatcp;
+use itertools;
 use light_curve_feature::{self as lcf, DataSample, FeatureEvaluator};
-use numpy::{IntoPyArray, PyArray1};
+use numpy::IntoPyArray;
 use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
 use pyo3::types::PyTuple;
 use std::convert::TryInto;
-
-type F = f64;
-type Feature = lcf::Feature<F>;
 
 const ATTRIBUTES_DOC: &str = r#"Attributes
 ----------
@@ -54,7 +52,8 @@ const COMMON_FEATURE_DOC: &str = formatcp!("\n{}\n\n{}\n", ATTRIBUTES_DOC, METHO
 
 #[pyclass(subclass, name = "_FeatureEvaluator")]
 pub struct PyFeatureEvaluator {
-    feature_evaluator: Feature,
+    feature_evaluator_f32: lcf::Feature<f32>,
+    feature_evaluator_f64: lcf::Feature<f64>,
 }
 
 impl PyFeatureEvaluator {
@@ -156,8 +155,8 @@ impl PyFeatureEvaluator {
         fill_value: Option<f64>,
     ) -> Res<PyObject> {
         let is_t_required = match (
-            self.feature_evaluator.is_t_required(),
-            self.feature_evaluator.is_sorting_required(),
+            self.feature_evaluator_f64.is_t_required(),
+            self.feature_evaluator_f64.is_sorting_required(),
             sorted,
         ) {
             // feature requires t
@@ -171,6 +170,28 @@ impl PyFeatureEvaluator {
         };
 
         match (t, m) {
+            (GenericFloatArray1::Float32(t), GenericFloatArray1::Float32(m)) => {
+                let sigma = sigma
+                    .map(|sigma| {
+                        sigma.try_into().map_err(|_| {
+                            Exception::ValueError(
+                                "sigma is float64, but t & m are float32".to_string(),
+                            )
+                        })
+                    })
+                    .map_or(Ok(None), |result| result.map(Some))?;
+                Ok(Self::call(
+                    &self.feature_evaluator_f32,
+                    t,
+                    m,
+                    sigma,
+                    sorted,
+                    is_t_required,
+                    fill_value.map(|v| v as f32),
+                )?
+                .into_pyarray(py)
+                .into_py(py))
+            }
             (GenericFloatArray1::Float64(t), GenericFloatArray1::Float64(m)) => {
                 let sigma = sigma
                     .map(|sigma| {
@@ -182,7 +203,7 @@ impl PyFeatureEvaluator {
                     })
                     .map_or(Ok(None), |result| result.map(Some))?;
                 Ok(Self::call(
-                    &self.feature_evaluator,
+                    &self.feature_evaluator_f64,
                     t,
                     m,
                     sigma,
@@ -194,7 +215,7 @@ impl PyFeatureEvaluator {
                 .into_py(py))
             }
             _ => Err(Exception::NotImplementedError(
-                "the supported only dtype is float64".into(),
+                "t and m have different dtype".into(),
             )),
         }
     }
@@ -202,13 +223,13 @@ impl PyFeatureEvaluator {
     /// Feature names
     #[getter]
     fn names(&self) -> Vec<&str> {
-        self.feature_evaluator.get_names()
+        self.feature_evaluator_f64.get_names()
     }
 
     /// Feature descriptions
     #[getter]
     fn descriptions(&self) -> Vec<&str> {
-        self.feature_evaluator.get_descriptions()
+        self.feature_evaluator_f64.get_descriptions()
     }
 }
 
@@ -221,17 +242,22 @@ impl Extractor {
     #[new]
     #[args(args = "*")]
     fn __new__(args: &PyTuple) -> PyResult<(Self, PyFeatureEvaluator)> {
-        let evals = args
-            .iter()
-            .map(|arg| {
-                arg.downcast::<PyCell<PyFeatureEvaluator>>()
-                    .map(|fe| fe.borrow().feature_evaluator.clone())
+        let evals_iter = args.iter().map(|arg| {
+            arg.downcast::<PyCell<PyFeatureEvaluator>>().map(|fe| {
+                let fe = fe.borrow();
+                (
+                    fe.feature_evaluator_f32.clone(),
+                    fe.feature_evaluator_f64.clone(),
+                )
             })
-            .collect::<Result<Vec<_>, _>>()?;
+        });
+        let (evals_f32, evals_f64) =
+            itertools::process_results(evals_iter, |iter| iter.unzip::<_, _, Vec<_>, Vec<_>>())?;
         Ok((
             Self {},
             PyFeatureEvaluator {
-                feature_evaluator: lcf::FeatureExtractor::new(evals).into(),
+                feature_evaluator_f32: lcf::FeatureExtractor::new(evals_f32).into(),
+                feature_evaluator_f64: lcf::FeatureExtractor::new(evals_f64).into(),
             },
         ))
     }
@@ -247,7 +273,7 @@ Parameters
     Feature objects
 {}
 "#,
-            lcf::FeatureExtractor::<F, Feature>::doc().trim_start(),
+            lcf::FeatureExtractor::<f64, lcf::Feature<f64>>::doc().trim_start(),
             COMMON_FEATURE_DOC,
         )
     }
@@ -266,7 +292,8 @@ macro_rules! evaluator {
                 (
                     Self {},
                     PyFeatureEvaluator {
-                        feature_evaluator: <$eval>::new().into(),
+                        feature_evaluator_f32: <$eval>::new().into(),
+                        feature_evaluator_f64: <$eval>::new().into(),
                     },
                 )
             }
@@ -310,6 +337,19 @@ macro_rules! fit_evaluator {
             }
         }
 
+        impl $name {
+            fn model_impl<T>(
+                t: Arr<T>,
+                params: Arr<T>,
+            ) -> ndarray::Array1<T>
+            where T: lcf::Float + numpy::Element
+            {
+                let params = ContCowArray::from_view(params.as_array(), true);
+                t.as_array()
+                    .mapv(|x| <$eval>::f(x, params.as_slice()))
+            }
+        }
+
         #[pymethods]
         impl $name {
             #[new]
@@ -343,27 +383,27 @@ macro_rules! fit_evaluator {
                     }
                 };
 
-                let eval = <$eval>::new(curve_fit_algorithm);
-
                 Ok((
                     Self {},
                     PyFeatureEvaluator {
-                        feature_evaluator: eval.into(),
+                        feature_evaluator_f32: <$eval>::new(curve_fit_algorithm.clone()).into(),
+                        feature_evaluator_f64: <$eval>::new(curve_fit_algorithm).into(),
                     },
                 ))
             }
 
             #[staticmethod]
             #[args(t, params)]
-            fn model<'py>(
-                py: Python<'py>,
-                t: Arr<'py, F>,
-                params: Arr<'py, F>,
-            ) -> &'py PyArray1<F> {
-                let params = ContCowArray::from_view(params.as_array(), true);
-                t.as_array()
-                    .mapv(|x| <$eval>::f(x, params.as_slice()))
-                    .into_pyarray(py)
+            fn model(
+                py: Python,
+                t: GenericFloatArray1,
+                params: GenericFloatArray1,
+            ) -> Res<PyObject> {
+                match (t, params) {
+                    (GenericFloatArray1::Float32(t), GenericFloatArray1::Float32(params)) => Ok(Self::model_impl(t, params).into_pyarray(py).into_py(py)),
+                    (GenericFloatArray1::Float64(t), GenericFloatArray1::Float64(params)) => Ok(Self::model_impl(t, params).into_pyarray(py).into_py(py)),
+                    _ => Err(Exception::ValueError("t and params must have the same dtype".to_string()))
+                }
             }
 
             #[classattr]
@@ -452,11 +492,12 @@ pub struct BeyondNStd {}
 impl BeyondNStd {
     #[new]
     #[args(nstd)]
-    fn __new__(nstd: F) -> (Self, PyFeatureEvaluator) {
+    fn __new__(nstd: f64) -> (Self, PyFeatureEvaluator) {
         (
             Self {},
             PyFeatureEvaluator {
-                feature_evaluator: lcf::BeyondNStd::new(nstd).into(),
+                feature_evaluator_f32: lcf::BeyondNStd::new(nstd as f32).into(),
+                feature_evaluator_f64: lcf::BeyondNStd::new(nstd).into(),
             },
         )
     }
@@ -471,7 +512,7 @@ Parameters
 nstd : positive float
     N
 {}"#,
-            lcf::BeyondNStd::<F>::doc().trim_start(),
+            lcf::BeyondNStd::<f64>::doc().trim_start(),
             COMMON_FEATURE_DOC,
         )
     }
@@ -490,24 +531,28 @@ impl Bins {
     fn __new__(
         py: Python,
         features: PyObject,
-        window: F,
-        offset: F,
+        window: f64,
+        offset: f64,
     ) -> PyResult<(Self, PyFeatureEvaluator)> {
-        let mut eval = lcf::Bins::default();
+        let mut eval_f32 = lcf::Bins::default();
+        let mut eval_f64 = lcf::Bins::default();
         for x in features.extract::<&PyAny>(py)?.iter()? {
-            let feature = x?
-                .downcast::<PyCell<PyFeatureEvaluator>>()?
-                .borrow()
-                .feature_evaluator
-                .clone();
-            eval.add_feature(feature);
+            let py_feature = x?.downcast::<PyCell<PyFeatureEvaluator>>()?.borrow();
+            eval_f32.add_feature(py_feature.feature_evaluator_f32.clone());
+            eval_f64.add_feature(py_feature.feature_evaluator_f64.clone());
         }
-        eval.set_window(window);
-        eval.set_offset(offset);
+
+        eval_f32.set_window(window as f32);
+        eval_f64.set_window(window);
+
+        eval_f32.set_offset(offset as f32);
+        eval_f64.set_offset(offset);
+
         Ok((
             Self {},
             PyFeatureEvaluator {
-                feature_evaluator: eval.into(),
+                feature_evaluator_f32: eval_f32.into(),
+                feature_evaluator_f64: eval_f64.into(),
             },
         ))
     }
@@ -526,7 +571,7 @@ window : positive float
 offset : float
     Zero time moment
 "#,
-            lcf::Bins::<F, Feature>::doc().trim_start()
+            lcf::Bins::<f64, lcf::Feature<f64>>::doc().trim_start()
         )
     }
 }
@@ -551,7 +596,8 @@ impl InterPercentileRange {
         (
             Self {},
             PyFeatureEvaluator {
-                feature_evaluator: lcf::InterPercentileRange::new(quantile).into(),
+                feature_evaluator_f32: lcf::InterPercentileRange::new(quantile).into(),
+                feature_evaluator_f64: lcf::InterPercentileRange::new(quantile).into(),
             },
         )
     }
@@ -603,7 +649,12 @@ impl MagnitudePercentageRatio {
         Ok((
             Self {},
             PyFeatureEvaluator {
-                feature_evaluator: lcf::MagnitudePercentageRatio::new(
+                feature_evaluator_f32: lcf::MagnitudePercentageRatio::new(
+                    quantile_numerator,
+                    quantile_denominator,
+                )
+                .into(),
+                feature_evaluator_f64: lcf::MagnitudePercentageRatio::new(
                     quantile_numerator,
                     quantile_denominator,
                 )
@@ -648,11 +699,13 @@ pub struct MedianBufferRangePercentage {}
 impl MedianBufferRangePercentage {
     #[new]
     #[args(quantile)]
-    fn __new__(quantile: F) -> (Self, PyFeatureEvaluator) {
+    fn __new__(quantile: f64) -> (Self, PyFeatureEvaluator) {
         (
             Self {},
             PyFeatureEvaluator {
-                feature_evaluator: lcf::MedianBufferRangePercentage::new(quantile).into(),
+                feature_evaluator_f32: lcf::MedianBufferRangePercentage::new(quantile as f32)
+                    .into(),
+                feature_evaluator_f64: lcf::MedianBufferRangePercentage::new(quantile).into(),
             },
         )
     }
@@ -667,7 +720,7 @@ Parameters
 quantile : positive float
     Relative range size      
 {}"#,
-            lcf::MedianBufferRangePercentage::<F>::doc(),
+            lcf::MedianBufferRangePercentage::<f64>::doc(),
             COMMON_FEATURE_DOC
         )
     }
@@ -687,7 +740,10 @@ impl PercentDifferenceMagnitudePercentile {
         (
             Self {},
             PyFeatureEvaluator {
-                feature_evaluator: lcf::PercentDifferenceMagnitudePercentile::new(quantile).into(),
+                feature_evaluator_f32: lcf::PercentDifferenceMagnitudePercentile::new(quantile)
+                    .into(),
+                feature_evaluator_f64: lcf::PercentDifferenceMagnitudePercentile::new(quantile)
+                    .into(),
             },
         )
     }
@@ -713,11 +769,12 @@ quantile : positive float
     text_signature = "(peaks=None, resolution=None, max_freq_factor=None, nyquist=None, fast=None, features=None)"
 )]
 pub struct Periodogram {
-    eval: lcf::Periodogram<F, Feature>,
+    eval_f32: lcf::Periodogram<f32, lcf::Feature<f32>>,
+    eval_f64: lcf::Periodogram<f64, lcf::Feature<f64>>,
 }
 
 impl Periodogram {
-    fn create_eval(
+    fn create_evals(
         py: Python,
         peaks: Option<usize>,
         resolution: Option<f32>,
@@ -725,16 +782,26 @@ impl Periodogram {
         nyquist: Option<PyObject>,
         fast: Option<bool>,
         features: Option<PyObject>,
-    ) -> PyResult<lcf::Periodogram<F, Feature>> {
-        let mut eval = match peaks {
+    ) -> PyResult<(
+        lcf::Periodogram<f32, lcf::Feature<f32>>,
+        lcf::Periodogram<f64, lcf::Feature<f64>>,
+    )> {
+        let mut eval_f32 = match peaks {
             Some(peaks) => lcf::Periodogram::new(peaks),
             None => lcf::Periodogram::default(),
         };
+        let mut eval_f64 = match peaks {
+            Some(peaks) => lcf::Periodogram::new(peaks),
+            None => lcf::Periodogram::default(),
+        };
+
         if let Some(resolution) = resolution {
-            eval.set_freq_resolution(resolution);
+            eval_f32.set_freq_resolution(resolution);
+            eval_f64.set_freq_resolution(resolution);
         }
         if let Some(max_freq_factor) = max_freq_factor {
-            eval.set_max_freq_factor(max_freq_factor);
+            eval_f32.set_max_freq_factor(max_freq_factor);
+            eval_f64.set_max_freq_factor(max_freq_factor);
         }
         if let Some(nyquist) = nyquist {
             let nyquist_freq: lcf::NyquistFreq =
@@ -753,26 +820,41 @@ impl Periodogram {
                         "nyquist must be one of: None, 'average', 'median' or quantile value",
                     ));
                 };
-            eval.set_nyquist(nyquist_freq);
+            eval_f32.set_nyquist(nyquist_freq.clone());
+            eval_f64.set_nyquist(nyquist_freq);
         }
         if let Some(fast) = fast {
             if fast {
-                eval.set_periodogram_algorithm(lcf::PeriodogramPowerFft::new().into());
+                eval_f32.set_periodogram_algorithm(lcf::PeriodogramPowerFft::new().into());
+                eval_f64.set_periodogram_algorithm(lcf::PeriodogramPowerFft::new().into());
             } else {
-                eval.set_periodogram_algorithm(lcf::PeriodogramPowerDirect {}.into());
+                eval_f32.set_periodogram_algorithm(lcf::PeriodogramPowerDirect {}.into());
+                eval_f64.set_periodogram_algorithm(lcf::PeriodogramPowerDirect {}.into());
             }
         }
         if let Some(features) = features {
             for x in features.extract::<&PyAny>(py)?.iter()? {
-                let feature = x?
-                    .downcast::<PyCell<PyFeatureEvaluator>>()?
-                    .borrow()
-                    .feature_evaluator
-                    .clone();
-                eval.add_feature(feature);
+                let py_feature = x?.downcast::<PyCell<PyFeatureEvaluator>>()?.borrow();
+                eval_f32.add_feature(py_feature.feature_evaluator_f32.clone());
+                eval_f64.add_feature(py_feature.feature_evaluator_f64.clone());
             }
         }
-        Ok(eval)
+        Ok((eval_f32, eval_f64))
+    }
+
+    fn freq_power_impl<T>(
+        eval: &lcf::Periodogram<T, lcf::Feature<T>>,
+        t: Arr<T>,
+        m: Arr<T>,
+    ) -> (ndarray::Array1<T>, ndarray::Array1<T>)
+    where
+        T: lcf::Float + numpy::Element,
+    {
+        let t: DataSample<_> = t.as_array().into();
+        let m: DataSample<_> = m.as_array().into();
+        let mut ts = lcf::TimeSeries::new_without_weight(t, m);
+        let (freq, power) = eval.freq_power(&mut ts);
+        (freq.into(), power.into())
     }
 }
 
@@ -796,7 +878,7 @@ impl Periodogram {
         fast: Option<bool>,
         features: Option<PyObject>,
     ) -> PyResult<(Self, PyFeatureEvaluator)> {
-        let eval = Self::create_eval(
+        let (eval_f32, eval_f64) = Self::create_evals(
             py,
             peaks,
             resolution,
@@ -806,26 +888,44 @@ impl Periodogram {
             features,
         )?;
         Ok((
-            Self { eval: eval.clone() },
+            Self {
+                eval_f32: eval_f32.clone(),
+                eval_f64: eval_f64.clone(),
+            },
             PyFeatureEvaluator {
-                feature_evaluator: eval.into(),
+                feature_evaluator_f32: eval_f32.into(),
+                feature_evaluator_f64: eval_f64.into(),
             },
         ))
     }
 
     /// Angular frequencies and periodogram values
     #[pyo3(text_signature = "(t, m)")]
-    fn freq_power<'py>(
+    fn freq_power(
         &self,
-        py: Python<'py>,
-        t: Arr<F>,
-        m: Arr<F>,
-    ) -> (&'py PyArray1<F>, &'py PyArray1<F>) {
-        let t: DataSample<_> = t.as_array().into();
-        let m: DataSample<_> = m.as_array().into();
-        let mut ts = lcf::TimeSeries::new_without_weight(t, m);
-        let (freq, power) = self.eval.freq_power(&mut ts);
-        (freq.into_pyarray(py), power.into_pyarray(py))
+        py: Python,
+        t: GenericFloatArray1,
+        m: GenericFloatArray1,
+    ) -> Res<(PyObject, PyObject)> {
+        match (t, m) {
+            (GenericFloatArray1::Float32(t), GenericFloatArray1::Float32(m)) => {
+                let (freq, power) = Self::freq_power_impl(&self.eval_f32, t, m);
+                Ok((
+                    freq.into_pyarray(py).into_py(py),
+                    power.into_pyarray(py).into_py(py),
+                ))
+            }
+            (GenericFloatArray1::Float64(t), GenericFloatArray1::Float64(m)) => {
+                let (freq, power) = Self::freq_power_impl(&self.eval_f64, t, m);
+                Ok((
+                    freq.into_pyarray(py).into_py(py),
+                    power.into_pyarray(py).into_py(py),
+                ))
+            }
+            _ => Err(Exception::ValueError(
+                "t and m must have the same dtype".to_string(),
+            )),
+        }
     }
 
     #[classattr]
@@ -886,7 +986,7 @@ Examples
 >>> peaks = periodogram(t, m, sorted=True)[::2]
 >>> frequency, power = periodogram.freq_power(t, m) 
 "#,
-            intro = lcf::Periodogram::<F, Feature>::doc(),
+            intro = lcf::Periodogram::<f64, lcf::Feature<f64>>::doc(),
             common = ATTRIBUTES_DOC,
         )
     }