+++ light-curve-python/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-—
+- Support `dtype=np.float32` for feature extractors
 
 ### Changed